interfaces Product
