+++ src/main/kotlin/kr/hhplus/be/server/interfaces/product/ProductController.kt
@@ -1,13 +1,6 @@
-package kr.hhplus.be.server.controller.product
+package kr.hhplus.be.server.interfaces.product
 
-import kr.hhplus.be.server.controller.product.api.ProductApi
-import kr.hhplus.be.server.controller.product.dto.request.ProductCreateRequest
-import kr.hhplus.be.server.controller.product.dto.request.ProductUpdateRequest
-import kr.hhplus.be.server.controller.product.dto.response.ProductDetailResponse
-import kr.hhplus.be.server.controller.product.dto.response.ProductOptionResponse
-import kr.hhplus.be.server.controller.product.dto.response.ProductResponse
-import kr.hhplus.be.server.controller.product.dto.response.TopSellingProductResponse
-import kr.hhplus.be.server.controller.product.dto.response.TopSellingProductsResponse
+import kr.hhplus.be.server.interfaces.product.api.ProductApi
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.validation.annotation.Validated
@@ -20,33 +13,33 @@ import jakarta.validation.Valid
 @Validated
 class ProductController : ProductApi {
 
-    override fun getAllProducts(): ResponseEntity<List<ProductResponse>> {
+    override fun getAllProducts(): ResponseEntity<List<ProductResponse.Response>> {
         // Mock API response
         val now = LocalDateTime.now()
         val products = listOf(
-            ProductResponse(
+            ProductResponse.Response(
                 productId = 1L,
                 name = "청바지",
                 description = "편안한 청바지입니다.",
                 price = 29000.0,
                 options = listOf(
-                    ProductOptionResponse(1, "S 사이즈", 0.0, 10),
-                    ProductOptionResponse(2, "M 사이즈", 0.0, 20),
-                    ProductOptionResponse(3, "L 사이즈", 0.0, 15)
+                    ProductResponse.ProductOptionResponse(1, "S 사이즈", 0.0, 10),
+                    ProductResponse.ProductOptionResponse(2, "M 사이즈", 0.0, 20),
+                    ProductResponse.ProductOptionResponse(3, "L 사이즈", 0.0, 15)
                 ),
                 createdAt = now,
                 updatedAt = now
             ),
-            ProductResponse(
+            ProductResponse.Response(
                 productId = 2L,
                 name = "티셔츠",
                 description = "시원한 여름용 티셔츠입니다.",
                 price = 15000.0,
                 options = listOf(
-                    ProductOptionResponse(4, "화이트 S", 0.0, 5),
-                    ProductOptionResponse(5, "화이트 M", 0.0, 10),
-                    ProductOptionResponse(6, "블랙 S", 2000.0, 3),
-                    ProductOptionResponse(7, "블랙 M", 2000.0, 8)
+                    ProductResponse.ProductOptionResponse(4, "화이트 S", 0.0, 5),
+                    ProductResponse.ProductOptionResponse(5, "화이트 M", 0.0, 10),
+                    ProductResponse.ProductOptionResponse(6, "블랙 S", 2000.0, 3),
+                    ProductResponse.ProductOptionResponse(7, "블랙 M", 2000.0, 8)
                 ),
                 createdAt = now,
                 updatedAt = now
@@ -55,18 +48,18 @@ class ProductController : ProductApi {
         return ResponseEntity.ok(products)
     }
 
-    override fun getProductById(@PathVariable id: Long): ResponseEntity<ProductDetailResponse> {
+    override fun getProductById(@PathVariable id: Long): ResponseEntity<ProductResponse.DetailResponse> {
         // Mock API response
         val now = LocalDateTime.now()
         val productId = id.toLong()
-        val product = ProductDetailResponse(
+        val product = ProductResponse.DetailResponse(
             productId = productId,
             name = "상품 $productId",
             description = "상품 상세 설명입니다.",
             price = 29000.0,
             options = listOf(
-                ProductOptionResponse(productId * 10 + 1, "옵션 1", 0.0, 10),
-                ProductOptionResponse(productId * 10 + 2, "옵션 2", 5000.0, 5)
+                ProductResponse.ProductOptionResponse(productId * 10 + 1, "옵션 1", 0.0, 10),
+                ProductResponse.ProductOptionResponse(productId * 10 + 2, "옵션 2", 5000.0, 5)
             ),
             createdAt = now,
             updatedAt = now
@@ -74,14 +67,14 @@ class ProductController : ProductApi {
         return ResponseEntity.ok(product)
     }
 
-    override fun getTopSellingProducts(): ResponseEntity<TopSellingProductsResponse> {
+    override fun getTopSellingProducts(): ResponseEntity<ProductResponse.TopSellingProductsResponse> {
         // 현재 시간 기준 최근 3일을 계산
         val now = LocalDateTime.now()
         val threeDaysAgo = now.minusDays(3)
         
         // Mock API response - 실제로는 OrderRepository를 통해 데이터를 가져와야 함
         val topSellingProducts = listOf(
-            TopSellingProductResponse(
+            ProductResponse.TopSellingProductResponse(
                 productId = 1L,
                 name = "청바지",
                 description = "편안한 청바지입니다.",
@@ -90,7 +83,7 @@ class ProductController : ProductApi {
                 createdAt = now.minusDays(10),
                 updatedAt = now.minusDays(1)
             ),
-            TopSellingProductResponse(
+            ProductResponse.TopSellingProductResponse(
                 productId = 5L,
                 name = "가죽자켓",
                 description = "고급 가죽 재질의 자켓입니다.",
@@ -99,7 +92,7 @@ class ProductController : ProductApi {
                 createdAt = now.minusDays(15),
                 updatedAt = now.minusDays(2)
             ),
-            TopSellingProductResponse(
+            ProductResponse.TopSellingProductResponse(
                 productId = 2L,
                 name = "티셔츠",
                 description = "시원한 여름용 티셔츠입니다.",
@@ -108,7 +101,7 @@ class ProductController : ProductApi {
                 createdAt = now.minusDays(20),
                 updatedAt = now.minusDays(3)
             ),
-            TopSellingProductResponse(
+            ProductResponse.TopSellingProductResponse(
                 productId = 8L,
                 name = "운동화",
                 description = "편안한 운동화입니다.",
@@ -117,7 +110,7 @@ class ProductController : ProductApi {
                 createdAt = now.minusDays(12),
                 updatedAt = now.minusDays(1)
             ),
-            TopSellingProductResponse(
+            ProductResponse.TopSellingProductResponse(
                 productId = 3L,
                 name = "모자",
                 description = "스타일리시한 모자입니다.",
@@ -129,7 +122,7 @@ class ProductController : ProductApi {
         )
         
         return ResponseEntity.ok(
-            TopSellingProductsResponse(
+            ProductResponse.TopSellingProductsResponse(
                 products = topSellingProducts,
                 startDate = threeDaysAgo,
                 endDate = now
@@ -137,16 +130,16 @@ class ProductController : ProductApi {
         )
     }
 
-    override fun createProduct(@Valid @RequestBody request: ProductCreateRequest): ResponseEntity<ProductResponse> {
+    override fun createProduct(@Valid @RequestBody request: ProductRequest.CreateRequest): ResponseEntity<ProductResponse.Response> {
         // Mock API response
         val now = LocalDateTime.now()
-        val createdProduct = ProductResponse(
+        val createdProduct = ProductResponse.Response(
             productId = 1,
             name = request.name,
             description = request.description,
             price = request.price,
             options = request.options.mapIndexed { index, option ->
-                ProductOptionResponse(
+                ProductResponse.ProductOptionResponse(
                     optionId = (index + 1).toLong(),
                     name = option.name,
                     additionalPrice = option.additionalPrice,
@@ -161,8 +154,8 @@ class ProductController : ProductApi {
 
     override fun updateProduct(
         @PathVariable id: Long,
-        @Valid @RequestBody request: ProductUpdateRequest
-    ): ResponseEntity<ProductResponse> {
+        @Valid @RequestBody request: ProductRequest.UpdateRequest
+    ): ResponseEntity<ProductResponse.Response> {
         // Mock API response
         val now = LocalDateTime.now()
         val productId = id.toLong()
@@ -174,18 +167,18 @@ class ProductController : ProductApi {
         
         // 옵션 처리
         val options = request.options?.map {
-            ProductOptionResponse(
+            ProductResponse.ProductOptionResponse(
                 optionId = it.optionId.toLong(),
                 name = it.name ?: "옵션 ${it.optionId}",
                 additionalPrice = it.additionalPrice ?: 0.0,
                 availableQuantity = it.availableQuantity ?: 10
             )
         } ?: listOf(
-            ProductOptionResponse(productId * 10 + 1, "기본 옵션 1", 0.0, 10),
-            ProductOptionResponse(productId * 10 + 2, "기본 옵션 2", 0.0, 5)
+            ProductResponse.ProductOptionResponse(productId * 10 + 1, "기본 옵션 1", 0.0, 10),
+            ProductResponse.ProductOptionResponse(productId * 10 + 2, "기본 옵션 2", 0.0, 5)
         )
         
-        val updatedProduct = ProductResponse(
+        val updatedProduct = ProductResponse.Response(
             productId = productId,
             name = name,
             description = description,
