+++ src/main/kotlin/kr/hhplus/be/server/interfaces/product/api/ProductApi.kt
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.controller.product.api
+package kr.hhplus.be.server.interfaces.product.api
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
@@ -7,11 +7,8 @@ import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.responses.ApiResponses
 import io.swagger.v3.oas.annotations.tags.Tag
-import kr.hhplus.be.server.controller.product.dto.request.ProductCreateRequest
-import kr.hhplus.be.server.controller.product.dto.request.ProductUpdateRequest
-import kr.hhplus.be.server.controller.product.dto.response.ProductDetailResponse
-import kr.hhplus.be.server.controller.product.dto.response.ProductResponse
-import kr.hhplus.be.server.controller.product.dto.response.TopSellingProductsResponse
+import kr.hhplus.be.server.interfaces.product.ProductRequest
+import kr.hhplus.be.server.interfaces.product.ProductResponse
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import jakarta.validation.Valid
@@ -24,18 +21,18 @@ interface ProductApi {
         ApiResponse(
             responseCode = "200",
             description = "조회 성공",
-            content = [Content(schema = Schema(implementation = ProductResponse::class))]
+            content = [Content(schema = Schema(implementation = ProductResponse.Response::class))]
         )
     )
     @GetMapping
-    fun getAllProducts(): ResponseEntity<List<ProductResponse>>
+    fun getAllProducts(): ResponseEntity<List<ProductResponse.Response>>
     
     @Operation(summary = "상품 상세 조회", description = "특정 상품의 상세 정보를 조회합니다.")
     @ApiResponses(
         ApiResponse(
             responseCode = "200",
             description = "조회 성공",
-            content = [Content(schema = Schema(implementation = ProductDetailResponse::class))]
+            content = [Content(schema = Schema(implementation = ProductResponse.DetailResponse::class))]
         ),
         ApiResponse(
             responseCode = "404",
@@ -47,25 +44,25 @@ interface ProductApi {
     fun getProductById(
         @Parameter(description = "상품 ID", required = true)
         @PathVariable id: Long
-    ): ResponseEntity<ProductDetailResponse>
+    ): ResponseEntity<ProductResponse.DetailResponse>
     
     @Operation(summary = "인기 상품 조회", description = "최근 3일 동안 가장 많이 팔린 상품 5개를 조회합니다.")
     @ApiResponses(
         ApiResponse(
             responseCode = "200",
             description = "조회 성공",
-            content = [Content(schema = Schema(implementation = TopSellingProductsResponse::class))]
+            content = [Content(schema = Schema(implementation = ProductResponse.TopSellingProductsResponse::class))]
         )
     )
     @GetMapping("/top-selling")
-    fun getTopSellingProducts(): ResponseEntity<TopSellingProductsResponse>
+    fun getTopSellingProducts(): ResponseEntity<ProductResponse.TopSellingProductsResponse>
     
     @Operation(summary = "상품 등록", description = "새로운 상품을 등록합니다.")
     @ApiResponses(
         ApiResponse(
             responseCode = "201",
             description = "등록 성공",
-            content = [Content(schema = Schema(implementation = ProductResponse::class))]
+            content = [Content(schema = Schema(implementation = ProductResponse.Response::class))]
         ),
         ApiResponse(
             responseCode = "400",
@@ -76,15 +73,15 @@ interface ProductApi {
     @PostMapping
     fun createProduct(
         @Parameter(description = "상품 생성 정보", required = true)
-        @Valid @RequestBody request: ProductCreateRequest
-    ): ResponseEntity<ProductResponse>
+        @Valid @RequestBody request: ProductRequest.CreateRequest
+    ): ResponseEntity<ProductResponse.Response>
     
     @Operation(summary = "상품 수정", description = "기존 상품 정보를 수정합니다.")
     @ApiResponses(
         ApiResponse(
             responseCode = "200",
             description = "수정 성공",
-            content = [Content(schema = Schema(implementation = ProductResponse::class))]
+            content = [Content(schema = Schema(implementation = ProductResponse.Response::class))]
         ),
         ApiResponse(
             responseCode = "400",
@@ -103,8 +100,8 @@ interface ProductApi {
         @PathVariable id: Long,
         
         @Parameter(description = "상품 수정 정보", required = true)
-        @Valid @RequestBody request: ProductUpdateRequest
-    ): ResponseEntity<ProductResponse>
+        @Valid @RequestBody request: ProductRequest.UpdateRequest
+    ): ResponseEntity<ProductResponse.Response>
     
     @Operation(summary = "상품 삭제", description = "상품을 삭제합니다.")
     @ApiResponses(
